+++ packages/samples/headless-ssr/src/app/generic/components/search-page.tsx
@@ -6,8 +6,8 @@ import {
   hydrateInitialState,
 } from '@/src/app/generic/common/engine';
 import {useEffect, useState} from 'react';
+import {HydrationMetadata} from '../../common/hydration-metadata';
 import {useSyncSearchParameters} from '../hooks/search-parameters';
-import {HydrationMetadata} from './hydration-metadata';
 import {ResultList} from './result-list';
 import {SearchBox} from './search-box';
 
@@ -36,7 +36,6 @@ export default function SearchPage({ssrState}: {ssrState: SearchSSRState}) {
 
   return (
     <>
-      <HydrationMetadata ssrState={ssrState} csrResult={csrResult} />
       <SearchBox
         ssrState={ssrState.controllers.searchBox.state}
         controller={csrResult?.controllers.searchBox}
@@ -45,6 +44,7 @@ export default function SearchPage({ssrState}: {ssrState: SearchSSRState}) {
         ssrState={ssrState.controllers.resultList.state}
         controller={csrResult?.controllers.resultList}
       />
+      <HydrationMetadata ssrState={ssrState} csrResult={csrResult} />
     </>
   );
 }