+++ packages/atomic/CHANGELOG.md
@@ -1,3 +1,20 @@
+# 2.43.0 (2023-09-11)
+
+### Bug Fixes
+
+- **atomic:** added part to access ::after ([#3151](https://github.com/coveo/ui-kit/issues/3151)) ([9b960e5](https://github.com/coveo/ui-kit/commits/9b960e585a2d52832d866d31a6e6150e8a18b03f))
+
+### Features
+
+- **commerce:** create PLP v2 interactive result controller ([#3129](https://github.com/coveo/ui-kit/issues/3129)) ([4d3c39d](https://github.com/coveo/ui-kit/commits/4d3c39d8a1b9910f528be0ed9c372c92e242ff9e))
+- **headless-react:** add headless react util defineSearchEngine ([#3107](https://github.com/coveo/ui-kit/issues/3107)) ([1d63e43](https://github.com/coveo/ui-kit/commits/1d63e431c427de9022238216cc5128b237b836b4))
+
+## 2.42.3 (2023-09-06)
+
+### Bug Fixes
+
+- **atomic:** translation issue for zh-tw ([5ac74a9](https://github.com/coveo/ui-kit/commits/5ac74a9e344c157057d2f99c6ab31db0c618554f))
+
 ## 2.42.1 (2023-08-30)
 
 ### Bug Fixes