+++ packages/headless/CHANGELOG.md
@@ -1,3 +1,18 @@
+# 2.31.0 (2023-09-11)
+
+### Features
+
+- **commerce:** create basic boilerplate for commerce sub-package [CAPI-82] ([#3125](https://github.com/coveo/ui-kit/issues/3125)) ([11c216d](https://github.com/coveo/ui-kit/commits/11c216d5ae9894b1d8cca260a96f68dea2e7c526))
+- **commerce:** create PLP v2 interactive result controller ([#3129](https://github.com/coveo/ui-kit/issues/3129)) ([4d3c39d](https://github.com/coveo/ui-kit/commits/4d3c39d8a1b9910f528be0ed9c372c92e242ff9e))
+- **headless-react:** add headless react util defineSearchEngine ([#3107](https://github.com/coveo/ui-kit/issues/3107)) ([1d63e43](https://github.com/coveo/ui-kit/commits/1d63e431c427de9022238216cc5128b237b836b4))
+- **headless:** add excludeFacetSearchResult search action ([#3149](https://github.com/coveo/ui-kit/issues/3149)) ([02480f6](https://github.com/coveo/ui-kit/commits/02480f6e97e54ad97974a8a93e8d784092e06973))
+
+## 2.30.1 (2023-09-06)
+
+### Features
+
+- **headless:** added defineSearchParametersManager SSR controller ([#3117](https://github.com/coveo/ui-kit/issues/3117)) ([99cca68](https://github.com/coveo/ui-kit/commits/99cca6864d94cbd53b9f60d4b049976b6f6a6981))
+
 # 2.29.0 (2023-08-30)
 
 ### Bug Fixes